+++ imports/components/@primitives/players/audio/FullPlayer.js
@@ -1,5 +1,6 @@
 
-import { Component, PropTypes } from "react";
+import PropTypes from "prop-types";
+import { Component } from "react";
 import { connect } from "react-redux";
 import { css } from "aphrodite";
 import collections from "../../../../util/collections";
@@ -51,10 +52,12 @@ class FullPlayerWithoutData extends Component {
   }
 
   setArtworkState = () => {
-    const artworkContainer = this.refs.artworkContainer;
-    this.setState({
-      isShort: artworkContainer.clientWidth > artworkContainer.clientHeight,
-    });
+    const artworkContainer = this.refs.artworkContainer || {};
+    if (artworkContainer.clientWidth && artworkContainer.clientHeight) {
+      this.setState({
+        isShort: artworkContainer.clientWidth > artworkContainer.clientHeight,
+      });
+    }
     return null;
   }
 
@@ -71,7 +74,7 @@ class FullPlayerWithoutData extends Component {
   // };
 
 
-  getArtworkStyles = (album) => {
+  getArtworkStyles = album => {
     const artworkContainer = this.refs.artworkContainer;
     const awStyles = backgrounds.styles(album);
 
@@ -84,7 +87,7 @@ class FullPlayerWithoutData extends Component {
     return awStyles;
   };
 
-  getArtworkClasses = (album) => {
+  getArtworkClasses = album => {
     const classes = [
       "one-whole",
       "overlay--gradient",
@@ -147,7 +150,6 @@ class FullPlayerWithoutData extends Component {
 
     if (state === "playing") {
       this.props.pause();
-      return;
     }
   }