+++ imports/pages/give/history/Filter.js
@@ -1,5 +1,7 @@
 // @flow
-import { Component, PropTypes } from "react";
+import PropTypes from "prop-types";
+
+import { Component } from "react";
 import moment from "moment";
 
 import TagSelect from "../../../components/@primitives/UI/forms/TagSelect";
@@ -51,7 +53,7 @@ export default class Filter extends Component {
 
   componentWillReceiveProps(nextProps: Object) {
     if (this.props.family.length !== nextProps.family.length) {
-      this.setState({ people: nextProps.family.map((x) => x.person.id) });
+      this.setState({ people: nextProps.family.map(x => x.person.id) });
     }
   }